Update drawer.md
Added a more information on what to do when you want to keep the default animation property on the 'new kendo.mobile.Application()' but want to use the drawer for navigating to some remote views without having the side effect of a double animation.  This is kind of explaining and giving an example of what the **Important** message above the addition.